requirements.txt
pillow
ultralytics
gradio
